+++ common/src/main/java/fr/loudo/narrativecraft/editors/cameraangle/CameraAngleMakerEditorMaker.java
@@ -378,7 +378,6 @@ public class CameraAngleMakerEditorMaker implements EditorMaker {
             entity.remove(Entity.RemovalReason.DISCARDED);
         }
         characterPlacements.removeIf(p -> p.getId().equals(placementId));
-        cameraAngle.getCharacterPlacements().removeIf(p -> p.getId().equals(placementId));
     }
 
     public void removeTemplateReference(UUID templateReferenceId) {
@@ -388,7 +387,6 @@ public class CameraAngleMakerEditorMaker implements EditorMaker {
                 removePlacement(placementId);
             }
         }
-        cameraAngle.getTemplateReferences().removeIf(ref -> ref.id().equals(templateReferenceId));
     }
 
     public List<Entity> getEntities() {